move some tests to test_ramp
